+++ Software/Install MS Edge.ps1
@@ -33,7 +33,7 @@ $BaseLogDir = "$PSScriptRoot\_Install Logs"       # [edit] add the location of y
 $PackageName = "$Product" 		            # [edit] enter the display name of the software (e.g. 'Arcobat Reader' or 'Microsoft Office')
 
 # Global variables
-$StartDir = $PSScriptRoot # the directory path of the script currently being executed
+# $StartDir = $PSScriptRoot # the directory path of the script currently being executed
 $LogDir = (Join-Path $BaseLogDir $PackageName)
 $LogFileName = ("$ENV:COMPUTERNAME - $PackageName.log")
 $LogFile = Join-path $LogDir $LogFileName
@@ -98,38 +98,43 @@ IF (Test-Path -Path "$PSScriptRoot\$Product\Version.txt") {
 	DS_WriteLog "I" "Installing $Product" $LogFile
 	try {
 		"$PSScriptRoot\$Product\MicrosoftEdgeEnterpriseX64.msi" | Install-MSIFile
+		DS_WriteLog "-" "" $LogFile
+		write-Host -ForegroundColor Green "...ready"
+		Write-Output ""
 		} catch {
-	DS_WriteLog "E" "Error installing $Product (error: $($Error[0]))" $LogFile       
+			DS_WriteLog "-" "" $LogFile
+			DS_WriteLog "E" "Error installing $Product (Error: $($Error[0]))" $LogFile
+			Write-Host -ForegroundColor Red "Error installing $Product (Error: $($Error[0]))"
+			Write-Output ""    
+			}
 	}
-	DS_WriteLog "-" "" $LogFile
 
-	# Disable scheduled tasks
-	Start-Sleep -s 5
-	$EdgeTasks= (Get-ScheduledTask | Where-Object {$_.TaskName -like "MicrosoftEdge*"}).TaskName
-	foreach ($Task in $EdgeTasks) {
-		Disable-ScheduledTask -TaskName $Task -EA SilentlyContinue | Out-Null
-		} 
-
-	# Disable Active Setup
-	$EdgeKey = (Get-Childitem -recurse "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components" | Get-Itemproperty | Where-Object { $_  -match 'Edge' }).PSChildName | Select-Object -First 1
-	Remove-Item -Path "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\$EdgeKey" -EA SilentlyContinue
-
-	# Disable Citrix API Hooks (MS Edge) on Citrix VDA
-	$(
-	$RegPath = "HKLM:SYSTEM\CurrentControlSet\services\CtxUvi"
-	IF (Test-Path $RegPath) {
-	$RegName = "UviProcessExcludes"
-	$EdgeRegvalue = "msedge.exe"
-	# Get current values in UviProcessExcludes
-	$CurrentValues = Get-ItemProperty -Path $RegPath | Select-Object -ExpandProperty $RegName
-	# Add the msedge.exe value to existing values in UviProcessExcludes
-	IF ($CurrentValues -notlike "*msedge.exe*") {
-		Set-ItemProperty -Path $RegPath -Name $RegName -Value "$CurrentValues$EdgeRegvalue;"
+	IF ((Get-ItemProperty HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {$_.DisplayName -eq "Microsoft Edge"} | Select-Object -First 1).DisplayVersion) {
+		# Disable scheduled tasks
+		Start-Sleep -s 5
+		$EdgeTasks= (Get-ScheduledTask | Where-Object {$_.TaskName -like "MicrosoftEdge*"}).TaskName
+		foreach ($Task in $EdgeTasks) {
+			Disable-ScheduledTask -TaskName $Task -EA SilentlyContinue | Out-Null
+			}
+		# Disable Active Setup
+		$EdgeKey = (Get-Childitem -recurse "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components" | Get-Itemproperty | Where-Object { $_  -match 'Edge' }).PSChildName | Select-Object -First 1
+		IF ($EdgeKey) {
+			Remove-Item -Path "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\$EdgeKey" -EA SilentlyContinue
+			}
+		# Disable Citrix API Hooks (MS Edge) on Citrix VDA
+		$(
+		$RegPath = "HKLM:SYSTEM\CurrentControlSet\services\CtxUvi"
+		IF (Test-Path $RegPath) {
+		$RegName = "UviProcessExcludes"
+		$EdgeRegvalue = "msedge.exe"
+		# Get current values in UviProcessExcludes
+		$CurrentValues = Get-ItemProperty -Path $RegPath | Select-Object -ExpandProperty $RegName
+		# Add the msedge.exe value to existing values in UviProcessExcludes
+		IF ($CurrentValues -notlike "*msedge.exe*") {
+			Set-ItemProperty -Path $RegPath -Name $RegName -Value "$CurrentValues$EdgeRegvalue;"
+			}
 		}
-	}
-	) | Out-Null
-	write-Host -ForegroundColor Green "...ready"
-	Write-Output ""
+		) | Out-Null
 	}
 
 	# Stop, if no new version is available
